update everything except the transactions page
