+++ app/kudoku/_graphql/query/getAllEWalletAccount.ts
@@ -1,0 +1,45 @@
+import client from '$utils/graphql';
+import { gql } from '@apollo/client';
+
+interface IGetAllEWalletAccount {
+  __typename: 'EWalletAccount';
+  institutionId: string;
+  accountNumber: string;
+  balance: string;
+}
+
+export async function getAllEWalletAccount(
+  token: string
+): Promise<IGetAllEWalletAccount[]> {
+  const query = gql`
+    query GetAllEWalletAccount {
+      getAllEWalletAccount {
+        institutionId
+        accountNumber
+        balance
+      }
+    }
+  `;
+
+  return new Promise((resolve, reject) => {
+    (async () => {
+      try {
+        const {
+          data: { getAllEWalletAccount },
+        } = await client.query({
+          query,
+          context: {
+            headers: {
+              Authorization: `Bearer ${token}`,
+            },
+          },
+          fetchPolicy: 'network-only',
+        });
+
+        resolve(getAllEWalletAccount);
+      } catch (e) {
+        reject(e);
+      }
+    })();
+  });
+}
