+++ app/kudoku/transaction/_graphql/query/getAllDebitAccount.ts
@@ -1,0 +1,41 @@
+import client from '$utils/graphql';
+import { gql } from '@apollo/client';
+
+interface IGetAllDebitAccount {
+  __typename: 'DebitAccount';
+  id: string;
+}
+
+export async function getAllDebitAccount(
+  token: string
+): Promise<IGetAllDebitAccount[]> {
+  const query = gql`
+    query GetAllDebitAccount {
+      getAllDebitAccount {
+        id
+      }
+    }
+  `;
+
+  return new Promise((resolve, reject) => {
+    (async () => {
+      try {
+        const {
+          data: { getAllDebitAccount },
+        } = await client.query({
+          query,
+          context: {
+            headers: {
+              Authorization: `Bearer ${token}`,
+            },
+          },
+          fetchPolicy: 'network-only',
+        });
+
+        resolve(getAllDebitAccount);
+      } catch (e) {
+        reject(e);
+      }
+    })();
+  });
+}
